+++ src/main/java/com/renomad/minum/ConfigErrorMessage.java
@@ -10,69 +10,6 @@ final class ConfigErrorMessage {
         // making this private to be clearer it isn't supposed to be instantiated.
     }
 
-    private static String getDefaultConfig() {
-        return """
-                ###
-                ### The following was generated as a stop-gap measure because the
-                ### mandatory config file was missing.  Please see the documentation
-                ### for including a more proper configuration file.  This version
-                ### lacks some fine-tuning keys.
-                
-                
-                ###  The port used for our plain, non-secure server
-
-                #SERVER_PORT=8080
-
-                #-----------------------------
-
-                ###  the port for our secure, TLS 1.3 server
-
-                #SECURE_SERVER_PORT=8443
-
-                #-----------------------------
-
-                ###  the name of our host on the internet.  Used by our
-                ###  system when it needs to build a self-referencing URL.
-
-                #HOST_NAME=localhost
-
-                ###  database top-level directory
-
-                #DB_DIRECTORY=out/simple_db
-
-                #-----------------------------
-
-                ###  The log levels are:
-                ###
-                ###  Related to the business purposes of the application.  That is,
-                ###  the very highest-level, perhaps least-common logging messages.
-                ###  AUDIT
-                ###
-                ###  Information useful for debugging.
-                ###  DEBUG
-                ###
-                ###
-                ###  Represents an error that occurs in a separate thread, so
-                ###  that we are not able to catch it bubbling up
-                ###  ASYNC_ERROR
-                ###
-                ###
-                ###  Information marked as trace is pretty much entered for
-                ###  the same reason as DEBUG - i.e. so we can see important
-                ###  information about the running state of the program. The
-                ###  only difference is that trace information is very voluminous.
-                ###  That is, there's tons of it, and it could make it harder
-                ###  to find the important information amongst a lot of noise.
-                ###  For that reason, TRACE is usually turned off.
-                ###  TRACE
-                ###
-                ###  list them here, separated by commas
-
-                #LOG_LEVELS=DEBUG,TRACE,ASYNC_ERROR,AUDIT
-
-                """.stripLeading();
-    }
-
     /**
      * If the user has no minum.config file in the location they
      * run the software, this message will be shown to assist. It
@@ -86,25 +23,16 @@ final class ConfigErrorMessage {
                 
                 
                 ----------------------------------------------------------------
-                ----------------- System misconfiguration ----------------------
+                ----------------- System Configuration Missing -----------------
                 ----------------------------------------------------------------
                 
                 No properties file found at ./minum.config
                 
-                A file named minum.config with the following contents must exist in
-                the directory you were at when running this program. Copy the
-                following text and save it as a file called minum.config at that
-                location.
+                Continuing, using defaults.  See source code for Minum for an
+                example minum.config, which will allow you to customize behavior.
                 
                 ----------------------------------------------------------------
                 ----------------------------------------------------------------
-                
-                    ****   Copy after this line -v    ****
-                
-                """ +
-                ConfigErrorMessage.getDefaultConfig() + """
-                
-                    ****   Copy before this line -^    ****
                 """;
     }
 }